requirements.txt
streamlit==1.24.0
opencv-python-headless==4.8.0
mediapipe==0.10.3
numpy==1.24.3
Pillow==9.5.0
